Converted many unsigned int and int to size_t where they represent the size of memory. This is necessary to build on 64-bit system with -Wshorten-64-to-32. Also updated a few printf format strings to have the 'l' modifier to fix printf format warnings from the compiler now that we are passing 64-bit values in those places. There are a few minor spots that have buffer lengths that I didn't bother converting (like the byte stack). All tests pass under 64-bit Mac OS X on 10.6.

+++ src/yajl_parser.c
@@ -46,9 +46,9 @@
 
 unsigned char *
 yajl_render_error_string(yajl_handle hand, const unsigned char * jsonText,
-                         unsigned int jsonTextLen, int verbose)
+                         size_t jsonTextLen, int verbose)
 {
-    unsigned int offset = hand->errorOffset;
+    size_t offset = hand->errorOffset;
     unsigned char * str;
     const char * errorType = NULL;
     const char * errorText = NULL;
@@ -66,7 +66,7 @@ yajl_render_error_string(yajl_handle hand, const unsigned char * jsonText,
     }
 
     {
-        unsigned int memneeded = 0;
+        size_t memneeded = 0;
         memneeded += strlen(errorType);
         memneeded += strlen(" error");
         if (errorText != NULL) {
@@ -87,8 +87,8 @@ yajl_render_error_string(yajl_handle hand, const unsigned char * jsonText,
     /* now we append as many spaces as needed to make sure the error
      * falls at char 41, if verbose was specified */
     if (verbose) {
-        unsigned int start, end, i;
-        unsigned int spacesNeeded;
+        size_t start, end, i;
+        size_t spacesNeeded;
 
         spacesNeeded = (offset < 30 ? 40 - offset : 10);
         start = (offset >= 30 ? offset - 30 : 0);
@@ -136,12 +136,12 @@ yajl_render_error_string(yajl_handle hand, const unsigned char * jsonText,
 
 
 yajl_status
-yajl_do_parse(yajl_handle hand, unsigned int * offset,
-              const unsigned char * jsonText, unsigned int jsonTextLen)
+yajl_do_parse(yajl_handle hand, size_t * offset,
+              const unsigned char * jsonText, size_t jsonTextLen)
 {
     yajl_tok tok;
     const unsigned char * buf;
-    unsigned int bufLen;
+    size_t bufLen;
 
   around_again:
     switch (yajl_bs_current(hand->stateStack)) {
